feat(react): filter-kit and page-header
