+++ packages/react/src/components/filter-kit/index.ts
@@ -1,0 +1,2 @@
+export { default as FilterPageHeader } from './FilterKit';
+export * from './FilterKit';
