--- packages/react/src/components/page-header/components/index.ts
@@ -1,2 +1,0 @@
-export { default as FilterPageHeader } from './FilterPageHeader';
-export * from './FilterPageHeader';
