+++ packages/react/src/components/page-header/index.ts
@@ -2,4 +2,3 @@ export { default as PageHeader } from './PageHeader';
 export { default as SubHeader } from './SubHeader';
 export * from './PageHeader';
 export * from './SubHeader';
-export * from './components';
